experiment preparation
